feat(video+desktop): camera capture, video UI, E2E AEAD wiring, test fixes
Blockers 4 & 5: browser getUserMedia → JPEG IPC → Rust I420 pipeline;
remote video strip renders decoded frames via canvas; EncryptingTransport
wraps QuinnTransport so WZP AEAD is applied to all media (C2 fix).

Test fixes: HandshakeResult.session destructuring across relay/client/crypto
integration tests; video_codecs field added to all CallOffer/CallAnswer
structs; wzp-video pipeline_roundtrip integration tests added.

PRD docs: five Kimi-ready specs for E2E encryption, Android NDK 0.9 migration,
quality upgrade flow, wire-format hardening, and clippy debt.

+++ crates/wzp-android/src/pipeline.rs
@@ -137,7 +137,7 @@ impl Pipeline {
         if header.fec_block != 0 {
             let is_repair = header.is_repair();
             if let Err(e) = self.fec_decoder.add_symbol(
-                header.fec_block as u8,
+                header.fec_block,
                 header.fec_block >> 8,
                 is_repair,
                 &packet.payload,
